+++ vue/components/ui/molecules/section-expandable/section-expandable.stories.js
@@ -1,5 +1,5 @@
 import { storiesOf } from "@storybook/vue";
-import { withKnobs, text, boolean } from "@storybook/addon-knobs";
+import { withKnobs, text, boolean, select } from "@storybook/addon-knobs";
 
 storiesOf("Components/Molecules/Section Expandable", module)
     .addDecorator(withKnobs)
@@ -13,6 +13,20 @@ storiesOf("Components/Molecules/Section Expandable", module)
             },
             animated: {
                 default: boolean("Animated", false)
+            },
+            uppercase: {
+                default: boolean("Uppercase", false)
+            },
+            size: {
+                default: select(
+                    "Size",
+                    {
+                        Small: "small",
+                        Medium: "medium",
+                        Large: "large"
+                    },
+                    "medium"
+                )
             }
         },
         data: function() {
@@ -30,6 +44,8 @@ storiesOf("Components/Molecules/Section Expandable", module)
                 v-bind:title="title"
                 v-bind:expanded="expandedData"
                 v-bind:animated="animated"
+                v-bind:uppercase="uppercase"
+                v-bind:size="size"
             >
                 <h3> This is some text inside the expandable section. </h3>
                 <p> And some smaller text. </p>