+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+* Added `iconProps`, `uppercase` and `size` to `section-expandable`'s props - [ripe-util-vue/#288](https://github.com/ripe-tech/ripe-util-vue/issues/288)
 * Add `pagination` component - [ripe-util-vue/#307](https://github.com/ripe-tech/ripe-util-vue/issues/307)
 * Add `primaryColor` and `secondaryColor` prop to `dropdown-button` - [ripe-util-vue/#288](https://github.com/ripe-tech/ripe-util-vue/issues/288)